rf: move s3 routes to Arsenal
Extracting routes for use in other S3-related projects.
