+++ lib/api/apiUtils/object/setUpCopyLocator.js
@@ -1,9 +1,8 @@
 const { errors } = require('arsenal');
-
 const { parseRange } = require('arsenal/lib/network/http/utils');
 
 const constants = require('../../../../constants');
-const routesUtils = require('../../../routes/routesUtils');
+const setPartRanges = require('./setPartRanges');
 /**
  * Uses the source object metadata and the requestHeaders
  * to determine the location of the data to be copied and the
@@ -59,7 +58,7 @@ function setUpCopyLocator(sourceObjMD, rangeHeader, log) {
                 };
         }
         if (range) {
-            dataLocator = routesUtils.setPartRanges(dataLocator, range);
+            dataLocator = setPartRanges(dataLocator, range);
             copyObjectSize = range[1] - range[0] + 1;
         }
     }
