+++ lib/api/objectGet.js
@@ -8,6 +8,7 @@ const collectResponseHeaders = require('../utilities/collectResponseHeaders');
 const validateHeaders = require('../utilities/validateHeaders');
 const { pushMetric } = require('../utapi/utilities');
 const { getVersionIdResHeader } = require('./apiUtils/object/versioning');
+const setPartRanges = require('./apiUtils/object/setPartRanges');
 const { metadataValidateBucketAndObj } = require('../metadata/metadataUtils');
 
 /**
@@ -118,6 +119,7 @@ function objectGet(authInfo, request, returnTagCount, log, callback) {
                         objMD['x-amz-server-side-encryption'];
                 }
             }
+            dataLocator = setPartRanges(dataLocator, byteRange);
         }
         pushMetric('getObject', log, {
             authInfo,
