+++ lib/server.js
@@ -4,9 +4,30 @@ const cluster = require('cluster');
 const arsenal = require('arsenal');
 
 const logger = require('./utilities/logger');
-const { clientCheck } = require('./utilities/healthcheckHandler');
+const { healthcheckHandler, clientCheck }
+    = require('./utilities/healthcheckHandler');
 const _config = require('./Config').config;
-const routes = require('./routes');
+const { unsupportedQueries, blacklistedPrefixes } = require('../constants');
+const api = require('./api/api');
+const data = require('./data/wrapper');
+const RedisClient = require('./RedisClient');
+const StatsClient = require('./StatsClient');
+
+const routes = arsenal.s3routes.routes;
+const allEndpoints = require('./utils').getAllEndpoints();
+const websiteEndpoints = _config.websiteEndpoints;
+
+// redis client
+let localCacheClient;
+if (_config.localCache) {
+    localCacheClient = new RedisClient(_config.localCache.host,
+        _config.localCache.port);
+}
+// stats client
+const STATS_INTERVAL = 5; // 5 seconds
+const STATS_EXPIRY = 30; // 30 seconds
+const statsClient = new StatsClient(localCacheClient, STATS_INTERVAL,
+    STATS_EXPIRY);
 
 class S3Server {
     /**
@@ -40,7 +61,17 @@ class S3Server {
     routeRequest(req, res) {
         // disable nagle algorithm
         req.socket.setNoDelay();
-        routes(req, res, logger);
+        const params = {
+            api,
+            healthcheckHandler,
+            statsClient,
+            allEndpoints,
+            websiteEndpoints,
+            blacklistedPrefixes,
+            unsupportedQueries,
+            dataRetrievalFn: data.get,
+        };
+        routes(req, res, params, logger);
     }
 
     /*
