+++ tests/unit/utils/responseStreamData.js
@@ -2,10 +2,10 @@ const assert = require('assert');
 const httpMocks = require('node-mocks-http');
 const { EventEmitter } = require('events');
 const { errors } = require('arsenal');
+const routesUtils = require('arsenal').s3routes.routesUtils;
 
 const { cleanup, DummyRequestLogger } = require('../helpers');
 const { ds } = require('../../../lib/data/in_memory/backend');
-const routesUtils = require('../../../lib/routes/routesUtils');
 const data = require('../../../lib/data/wrapper');
 
 const responseStreamData = routesUtils.responseStreamData;
@@ -46,7 +46,7 @@ describe('responseStreamData:', () => {
             done();
         });
         return responseStreamData(errCode, overrideHeaders,
-            resHeaders, dataLocations, response, null, log);
+            resHeaders, dataLocations, data.get, response, null, log);
     });
 
     it('should stream full requested object data for two part object', done => {
@@ -74,7 +74,7 @@ describe('responseStreamData:', () => {
             done();
         });
         return responseStreamData(errCode, overrideHeaders,
-            resHeaders, dataLocations, response, null, log);
+            resHeaders, dataLocations, data.get, response, null, log);
     });
 
     it('#334 non-regression test, destroy connection on error', done => {
@@ -102,6 +102,6 @@ describe('responseStreamData:', () => {
             done(new Error('end reached instead of destroying connection'));
         });
         return responseStreamData(errCode, overrideHeaders,
-            resHeaders, dataLocations, response, null, log);
+            resHeaders, dataLocations, data.get, response, null, log);
     });
 });
